Throw exception when a view has no height or width (#380)
This usually means that the test has moved on to another activity and the one passed in is no longer being displayed. This will hopefully help others who get here: instead of a RTE when Bitmap tries to create a bitmap with no height/width, we'll throw an exception with a useful message.