Fix skipping devices
Fixes bug where the `--skip-serial` arguments would only be applied if you also used the `--serial` argument. The devices serial filtering was moved from the `SpoonRunner` constructor to the `run()` method, so the filter can be applied correctly.

Partially reverts the changes in #480, but still makes sure the set of serials is mutable before removing skipped devices.