[bug] Coverage files getting overwritten for non --single-instrumentation-call Spoon task (#582)
* [bug]

The multi instrumented spoon task (where --single-instrumentation-call is not provided as argument) works by running command `adb shell am instrument` for each test case.

If the coverage flag is enabled for spoon task, it will create a coverage file for each test case. This behavior is not causing any problems for --single-instrumentation-call tasks. But for multi instrumented tasks, Spoon is replacing the previous test's coverage file when the next test runs because each file have the same name (`coverage.ec`).

This change solves the issue by pulling the file from device after each test runs and in the merging the all the *.ec files into one.

Coverage file for each test will be available to the user with a proper naming convention for debugging purposes.

* [refactor] Remove duplicate code in pullCoverageFile() method

* Add unit test for mergeAllCoverageFiles method

* Add method descriptions

* [fix] Travis CI Build Failure fix.

The original shell script was not accepting the licenses.
Fixed this using sdkmanager to accept all the licenses.